Closes #521 Implemented PhantomJS image diff testing
See the UIActorSpec.ts for an example of how to write an image diff test using phantomjs. Images for a particular spec are stored in src/spec/images/SpecName/image.png. This issue demonstrated some of the limitations of PhantomJS, some things flat out don't work like they should in the browser. Forcing the implementation of the ex.Engine.suppressMinimumFeatureDetection and ex.Engine.suppressConsoleBootMessage to mitigate. Suppressing minimum feature detection was necessary because object url feature testing was failing in phantomjs for unknown reasons, however the code works in phantomjs.