[#521] Implement Image Diff Testing in Excalibur (#681)
Closes #521

See the UIActorSpec.ts for an example of how to write an image diff test using PhantomJS. Images for a particular spec are stored in src/spec/images/SpecName/image.png. 

This issue demonstrated some of the limitations of PhantomJS. Some things don't work like they should in the browser. Forcing the implementation of the ex.Engine.suppressMinimumFeatureDetection and ex.Engine.suppressConsoleBootMessage to mitigate these issues. Suppressing minimum feature detection was necessary because object URL feature testing was failing in PhantomJS for unknown reasons; however, the code works in PhantomJS.

## Changes:
* Re-implement local debugging experience in vscode
* Update travis yml to support 4, fix peerDependency issue
* Fix appveyor node version
* Remove unneeded jasmine dependency
* Add nicer matcher interface, clean up typings file for imagediff