[#97] Add basic frame statistics (#694)
Part of #97

To make our way towards closing #97 we need basic statistics to gather and display. This adds some basic frame statistics that are collected during a frame and is designed to be extended later on with more stats. Since all stats are encapsulated in a class, users can add historical tracking and logging of past statistics using the `preframe` and `postframe` hooks on engine.

## Changes:

- Add `Engine.debug` to access current frame stats and (later) provide debugging flags/functions
- Add `Engine.stats.currFrame` and `Engine.stats.prevFrame` to hold current and previous frame statistics
- Add new `FrameStats` class to hold frame statistics
- Add new `postframe` and `preframe` events which fire before and after the frame is drawn/updated to aid in stat display/collection