[chore] [docs] Fix breakout ball collision step
If you're following along with the quickstart guide, after the step
where you add the ball, it says the ball will bounce off of the paddle
but it doesn't until you've added the precollision handler later on
in the tutorial.

This change adds the precollision handler when adding the ball, then
updates it later for breaking the bricks.