Optimize moving the active chunk window at start
Improve moving the "chunk window" to the currently used location without
having to allocate and generate all the chunks that are outside of the
viewport anyway.

This fixes support for large chunked maps which would result in a
memory overflow while trying to extend the current chunks-in-memory
matrix if the first frame focused on a part that is far enough from
the top-left corner of the map.

See also: #1348