Make chunk system generator's API easier to understand
Drop the chunkSize parameter, since it can be retrieved from the
chunkSystemTileMap parameter and there are already multiple properties that are
to be accessed this way in the generator.

Rename the chunkX and chunkY parameters to chunkColumn and chunkRow to make it
easier to understand their actual meaning, hoping their usage will be less
confusing.

See also: #1348