Add a demo of the chunk system tile map
Add a proof-of-concept demo of the ChunkSystemTileMap. This proof-of-concept is
used to catch a fix the most apparant bugs and asses performance of the current
solution. This demo will have its chunk generator replaced with a generator
that uses a Perlin noise generator in the future to turn it into a propper
demonstration.

See also: #1348