Utilize core-js polyfills to greater extend in favor of custom polyfills
This commit also reconfigures the TypeScript compiler's assumed libraries, and
updates the list of included polyfills to be compatible with the current
configuration of the TypeScript compiler. The list of included polyfills is
slighty larger than originally intended since some lib type definitions
(es2015.core -
https://github.com/microsoft/TypeScript/blob/master/lib/lib.es2015.core.d.ts)
are not very finely granular.

See also: #1348