Move Actor-TileMap/ChunkSystemTileMap collision logic to a single class
Refactor collision detection and resolution logic to a single shared trait,
including the collide method of ChunkSystemTileMap. The collide method of
TileMap is preserved in its current form to better preserve backwards
compatibility.

This change also deprecates the TileMapCollisionDetection trait and the
TileMap.collides() method, marking them for removal in v0.26.0 in favor of the
now centralized logic located in the CellMapCollisionDetection trait.

See also: #1348