Actor: Separate color and opacity.
Previously color and opacity were tied together in a confusing way.
Constructing an actor with a specified color would cause the opacity
to be set to the color's alpha value, but subsequent changes to the
color's alpha value would effectively be ignored in favour of opacity.

The Color's alpha value and Actor's opacity are now separate and both
will be honored by multiplying them together.