fix: Prevent pair generation for composite colliders (#2131)
Fixes an issue where pairs were erroneously being generated for composite colliders when fps was low.

<img width="768" alt="image" src="https://user-images.githubusercontent.com/612071/143729442-64926afd-1914-41b2-9f76-eb7202a2912e.png">

Discovered when artificially limiting to 15 fps, this happened because the low fps would trigger the continuous fast moving object detection for the collision processor. This fast object code had a bug and did not evaluate whether it should generate a pair the same way. The code has been fixed so that they are consistent.

## Changes:

- Prevent pairs with the same owner id from being generated
- Adds tests to the fast moving object