fix: [#2300] CompositeColliders count as a whole for collisionstart/collisionend events (#2301)
Closes #2300 

This PR solves the issue of duplicate `collisionstart`/`collisionend` events by keeping track of the `CompositeCollider` in the contact id (which makes it no longer unique) but allows the dispatch logic for start/end to work unchanged in the `CollisionSystem`. This is done with a slightly gross hack to decorate colliders with a composite id if they are part of a composite collider.