feat: [#2272] Add Fit Screen/Container and Zoom (#2312)
Related #2272

This PR implements a fit screen/container and zoom to avoid the letter boxing, this is achieved by over scaling the viewport and positioning. One caveat is this will clip some drawable area so should only be used in games that are okay with losing some draw space. This PR also adds a `ex.Screen.contentArea` which will return the "safe space" to place content given the current screen size.

New display modes `ex.DisplayMode.FitScreenAndZoom` and `ex.DisplayMode.FitContainerAndZoom`

![image](https://user-images.githubusercontent.com/612071/168433294-5bc3467a-f308-4e3f-9b78-30c0a3dce7b5.png)

Notice that some of the viewport is offscreen