feat: [#2335] Add current playback time & playback rate (#2336)
Closes #2335 

This PR adds a new method onto `ex.Sound.getCurrentTime()` to get the current playback time into the currently playing track. 

Additionally this PR adds the current playback rate that can be get/set on `ex.Sound.playbackRate`, 1.0 is the default playback rate, 2.0 is twice the speed, and 0.5 is half speed.