feat: QOL Update import site for ex.Input  (#2697)
This PR deprecates the old `ex.Input.*` import site in favor of just importing off of `ex.*`. 

This is a quality of life improvement, it is rather annoying to need to either import `Input` to use input types, or to type `ex.Input.SomeImport` to use a type.

The old import site will continue to work for now, you'll just see the deprecation warning in the editor.