fix: Adjust font cache to prevent downward spiral on mobile (#2850)
This PR fixes a downward spiral that is possible on platforms with a slow couple frames. The `FontCache` font timeout to 500 ms and makes it configurable as a static `FontCache.FONT_TIMEOUT`. This is to help prevent a downward spiral on mobile devices that might take a long while to render a few starting frames causing the cache to repeatedly clear and never recover.