chore: enforce prettier and eslint (#3014)
See issue #2378 

- Add prettier for code formatting
- Update eslint config to parse other parts of codebase outside of `src/engine`, and also remove rules that are handled by prettier
- Add husky & lint-staged to enforce formatting on git commit
- Add format step to `all` and `all:ci` scripts

I separated the commit that runs the format & lint on all files into #3015 so the config can be easily reviewed here. 

Once we're happy with this, we can merge that PR into this one and ignore the commit(s) using `git blame --ignore-rev` to remove it from git blame history.