feat: Add actor.oldGlobalPos, deprecate getGlobalPos in favour of globalPos getter (#3017)
Closes #2922

## Changes:

- `actor.oldGlobalPos` returns the globalPosition from the previous frame
- `actor.getGlobalPos()` - use `actor.globalPos` instead
- `actor.getGlobalRotation()` - use `actor.globalRotation` instead
- `actor.getGlobalScale()` - use `actor.globalScale` instead

This required some changes to `MotionSystem` so that `globalPos` could be properly captured. Previously, the parent's transform may have already been updated in the system's update when processing a child entity, leading to an incorrect `oldGlobalPos` value. Now, the MotionSystem update captures the entity & all of its children at the same time, skipping any entities that have parents as to not capture them twice over.

This also deprecates other `getGlobalXYZ()` functions and replaces them with getters for consistency.